Fix cpu util not showing

+++ src/processor.cpp
@@ -2,4 +2,4 @@
 #include "linux_parser.h"
 
 // TODO: Return the aggregate CPU utilization
-float Processor::Utilization() { return LinuxParser::ActiveJiffies() / LinuxParser::Jiffies(); }
+float Processor::Utilization() { return static_cast<float>(float(LinuxParser::ActiveJiffies()) / float(LinuxParser::Jiffies())); }
